+++ src/app/es-feedback/_components/EsForm/constants.ts
@@ -1,0 +1,46 @@
+export const dummyResult = [
+  {
+    input_text: 'dddddddddddd',
+    response: {
+      text1: {
+        添削前文章: 'dddddddddddd',
+        添削語文章: '（この文章は空白または不完全です）',
+        添削理由: '入力された文章は意味がなく、評価および添削ができません。',
+      },
+      text2: {
+        添削前文章: 'dddddddddddd',
+        添削語文章: '（この文章は空白または不完全です）',
+        添削理由:
+          '入力された文章は内容が明確でなく、評価項目に基づく添削が不可能です。',
+      },
+    },
+  },
+  {
+    input_text: 'dddddddddddd',
+    response: {
+      error: {
+        message:
+          '申し訳ありませんが、提供された内容が適したエントリーシートの文章ではないため、添削ができません。具体的な文章を送っていただければ、評価項目に基づいて添削を行います。',
+      },
+    },
+  },
+  {
+    input_text: 'ddddddddddd',
+    response: {
+      text1: {
+        添削前文章:
+          '指定されたエントリーシートの文章の添削に必要な具体的な文章が提供されていないため、添削を行うことができません。',
+        添削後文章:
+          'エントリーシートの添削には具体的な文章が必要ですので、提供をお願いします。',
+        添削理由: '文章を簡潔にし、依頼内容を明確にするため。',
+      },
+      text2: {
+        添削前文章:
+          'エントリーシートの具体的な文章を提供していただくか、例文を提示してください。',
+        添削後文章:
+          '具体的なエントリーシートの文章を提供するか、例文を提示してください。',
+        添削理由: '文章の主語を明示し、呼びかけの内容を明確にするため。',
+      },
+    },
+  },
+]
